pyproject.toml
[tool.poetry]
name = "release_manager"
version = "0.1.0"
description = ""
authors = ["menikolaev <devmenikolaev@gmail.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
python-dateutil = "^2.8.2"
GitPython = "^3.1.20"
pyTelegramBotAPI = "^4.0.0"

[tool.poetry.dev-dependencies]
requests = "^2.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
